Filter - Change species for a country, state, or county
Filters can now be applied for observations of a species within a
country, state, or county. Adding country and state is a little
speculative since there is no existing need. However adding a filter
for county is currently useful in the case of differentiating between
Feral Pigeons and 'wild' Rock Doves.