Removed the 'multiple' flag from Observer
A cleaner way of dealing with Observers with multiple names is to
prime the database with two Observer records for each name. That way
the APILoader will scrape the checklist web page to obtain the
observer identifier, e.g. USER3445322, and so the checklist will be
assigned to the correct observer.